Add model GoalTemplate and created a Dialog for adding a goal
Created goal_template model with GoalTemplate, StageTemplate and TaskTemplate classes and created a list that fills the models with real data for templates. Also created a dialog to add a goal and removed unused add_goal screen. Created a placeholder for choose_goal_template and linked the dialog with goRoute.